+++ drivers/video/tegra/nvmap/nvmap_fault.c
@@ -56,7 +56,7 @@ void nvmap_vma_open(struct vm_area_struct *vma)
 	struct list_head *tmp_head = NULL;
 	pid_t current_pid = task_tgid_nr(current);
 	bool vma_pos_found = false;
-	int nr_page, i;
+	size_t nr_page, i;
 	ulong vma_open_count;
 
 	priv = vma->vm_private_data;
@@ -134,7 +134,7 @@ static void nvmap_vma_close(struct vm_area_struct *vma)
 	struct nvmap_vma_list *vma_list;
 	struct nvmap_handle *h;
 	bool vma_found = false;
-	int nr_page, i;
+	size_t nr_page, i;
 
 	if (!priv)
 		return;
